social_network/requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachy==0.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cleo==0.8.1
clikit==0.6.2
crashtest==0.3.1
cryptography==3.4.8
distlib==0.3.4
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
filelock==3.7.0
html5lib==1.1
idna==3.3
jeepney==0.7.1
keyring==21.8.0
lockfile==0.12.2
msgpack==1.0.3
packaging==20.9
pastel==0.2.1
pexpect==4.8.0
pipenv==2022.5.2
pkginfo==1.8.2
platformdirs==2.5.2
poetry==1.1.13
poetry-core==1.0.8
ptyprocess==0.7.0
pycosat==0.6.3
pycparser==2.21
PyJWT==2.4.0
pylev==1.4.0
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
requests-toolbelt==0.9.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
shellingham==1.4.0
sqlparse==0.4.2
tomlkit==0.10.2
urllib3==1.26.9
virtualenv==20.14.1
virtualenv-clone==0.5.7
webencodings==0.5.1
